backend/requirements.txt
fastapi>=0.115
uvicorn>=0.34
sqlalchemy>=2.0
watchdog>=3.0
faster-whisper>=1.1
scenedetect[opencv]>=0.6.5
python-multipart
aiofiles
python-dotenv>=1.0
anthropic>=0.39
Pillow>=10.0
google-api-python-client>=2.0
google-auth>=2.0
google-auth-oauthlib>=1.0
google-auth-httplib2>=0.2
